NIST'23 benchmark: compat fixes + stage-1 fragment eval
Fixes uncovered running the full 4-model scaffold benchmark:
- numpy>=2: safe_assign uses np.int64 (np.integer is abstract, forbidden as a
  dtype) — unblocks ICEBERG/MARASON predict + inten and GLACIER.
- GLACIER joint_model: lr_scheduler_step(..., metric=None) for the PL 2.x hook
  signature; joint_train config num-workers 32 -> 4 (4-GPU DDP was spawning
  ~128 dataloader workers and exhausting host RAM).
- MARASON inten: strategy 'ddp_find_unused_parameters_true' (the RAG reference
  params are unused in the loss when reference augmentation is disabled).
- ICEBERG train_gen: build_gen_magma_map enumerates via PredSpecDB so both a
  single magma_tree.hdf5 and sharded magma_tree_shard*.hdf5 are handled.

New benchmark tooling:
- analysis/dag_frag_eval.py: stage-1 fragment-DAG P/R/F1/Jaccard on HDF5 trees
  (WL-hash fragment identity; --split-name/--subset to score test-only).
- run_scripts/nist23_benchmark/stage1_frag_eval.sh + RESULTS.md (final table).

+++ src/ms_pred/common/misc_utils.py
@@ -73,7 +73,7 @@ class MassSpec:
                 return None
             if not np.issubdtype(x_np.dtype, dtype):
                 if dtype == np.integer:
-                    x_np = x_np.astype(np.integer)
+                    x_np = x_np.astype(np.int64)  # np.integer is abstract; numpy>=2.0 forbids it as a dtype
                 else:
                     raise TypeError(f'Input data does not have the correct data type, expected {dtype}, got {x_np.dtype}')
             return x_np
